requirements.txt
requests
matplotlib
yahoofinance
tabulate
python-dateutil
yfinance
pdfplumber
